Move UI helper classes into 'controls' path
